More refactoring
More refactoring, moving methods from Default to GenericUtilsProtocol
